fix up

+++ src/main/java/com/t/medicaldocument/entity/Vo/UserVo.java
@@ -24,5 +24,5 @@ public class UserVo {
 	 */
 	private Integer userPoints;
 
-	private Long userCapacity;
+	private Integer userCapacity;
 }
